// api/api.ts
export interface AreaOut {
    /**
     * 
     * @type {string}
     * @memberof AreaOut
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof AreaOut
     */
    'id': string;
    /**
     * 
     * @type {LocationOut}
     * @memberof AreaOut
     */
    'location': LocationOut;
    /**
     * 
     * @type {UserSchema}
     * @memberof AreaOut
     */
    'user': UserSchema;
}
// ...
export interface LocationOut {
    /**
     * 
     * @type {string}
     * @memberof LocationOut
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof LocationOut
     */
    'id': string;
    /**
     * 
     * @type {UserSchema}
     * @memberof LocationOut
     */
    'user': UserSchema;
}
// ...
export interface PlantOut {
    /**
     * 
     * @type {string}
     * @memberof PlantOut
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof PlantOut
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof PlantOut
     */
    'common_name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlantOut
     */
    'scientific_name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlantOut
     */
    'purchase_date'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof PlantOut
     */
    'graveyard'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof PlantOut
     */
    'death_date'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlantOut
     */
    'main_photo'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlantOut
     */
    'notes'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlantOut
     */
    'area': string;
    /**
     * 
     * @type {string}
     * @memberof PlantOut
     */
    'user': string;
}
// ...
export interface UserSchema {
    /**
     * 
     * @type {string}
     * @memberof UserSchema
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof UserSchema
     */
    'first_name': string;
    /**
     * 
     * @type {string}
     * @memberof UserSchema
     */
    'last_name': string;
}